Updated map_crawl and moved to its own file.
Consolidated some functions and moved all the of the map_crawling to one
file.
