+++ includes/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 10:46:43 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/21 15:35:26 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/22 12:56:36 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 typedef struct s_list	t_list;
 
 /*
- *	0, 1, 2,
- *	7, C, 3,
- *	6, 5, 4
+ *	7, 6, 5,
+ *	0, C, 4,
+ *	1, 2, 3
  */
 
 typedef struct s_coord
